fix(build): UnitList_DrawPageHeader stale s8 local-decl broke clean build; + port EkrPrepareBanimfx
D153 regression: last iter's unitlistscreen.h proto s8->int conflicted with a
local forward-decl in unitlistscreen_0809404C.c (incremental build hid it; clean
build fails). Fixed the local decl. EkrPrepareBanimfx: u16->int + (s16) cast.

+++ include/ekrbattle.h
@@ -882,7 +882,7 @@ void BeginAnimsOnBattleAnimations(void);
 void EkrMainEndExec(void);
 void MainUpdate_0(void);
 
-void EkrPrepareBanimfx(struct Anim * anim, u16);
+void EkrPrepareBanimfx(struct Anim * anim, int);
 s16 GetEfxHp(int index);
 // ??? GetEfxHpModMaybe(???);
 u16 IsItemDisplayedInBattle(u16 item);
